Tools/ApplicationDebugger: update jacobi sample. (#1421)
* Restructure the source code:
  - the common parts are moved to jacobi.cpp;
  - the computation parts have two flavors:
    - bugged.cpp contains computation with bugs;
    - fixed.cpp contains fixed computation;
* Simplify reductions;
* Update README.
